Logs: remove duplicates
